Continued to tease apart native and driver indexes
